move core data code into a service and use a predicate to filter rather than for loop
