+++ amatracker/RecordService.swift
@@ -1,0 +1,43 @@
+//
+//  RecordService.swift
+//  amatracker
+//
+//  Created by ryan on 8/16/15.
+//  Copyright (c) 2015 ryan. All rights reserved.
+//
+
+import Foundation
+import CoreData
+import UIKit
+
+class RecordService {
+    
+    let appDelegate : AppDelegate = (UIApplication.sharedApplication().delegate as! AppDelegate)
+    
+    func loadRecords() -> [Record]! {
+        let request = NSFetchRequest(entityName: "Record")
+        
+        // add type since we're casting
+      
+        let context = appDelegate.managedObjectContext!
+        
+        var error: NSError?
+        let fetchedRecords = context.executeFetchRequest(request, error: &error) as! [Record]
+        
+        return fetchedRecords
+    }
+    
+    func loadRecordsByMovement(movement: Movement) -> [Record]! {
+        let movementPredicate = NSPredicate(format: "movement == %@", movement)
+        let request = NSFetchRequest(entityName: "Record")
+        request.predicate = movementPredicate
+        
+        let context = appDelegate.managedObjectContext!
+        
+        var error: NSError?
+        let fetchedRecords = context.executeFetchRequest(request, error: &error) as! [Record]
+        
+        return fetchedRecords
+        
+    }
+}
